+++ backends/qualcomm/tests/test_qnn_delegate.py
@@ -698,6 +698,17 @@ class TestQNNQuantizedOperator(TestQNN):
                 )
                 self.lower_module_and_test_output(module, sample_input)
 
+    def test_qnn_backend_16a4w_conv2d_qat(self):
+        modules = [Conv2dSingle(), Conv2dSingle(bias=False)]  # noqa: F405
+        sample_input = (torch.randn([1, 1, 3, 3]),)
+        for i, module in enumerate(modules):
+            with self.subTest(i=i):
+                prepared = self.get_prepared_qat_module(module, sample_input)
+                converted = self.get_converted_sgd_trained_module(
+                    module, prepared, sample_input
+                )
+                self.lower_module_and_test_output(converted, sample_input)
+
     def test_qnn_backend_16a4w_layer_norm(self):
         module = LayerNorm()  # noqa: F405
         sample_input = (torch.randn(196, 768),)
@@ -1063,18 +1074,8 @@ class TestQNNQuantizedOperator(TestQNN):
         """
         module = Linear()  # noqa: F405
         sample_input = (torch.randn([3, 4]),)
-
-        module = self.get_prepared_qat_module(module, sample_input)
-
-        optimizer = torch.optim.SGD(module.parameters(), lr=0.1)
-        criterion = torch.nn.CrossEntropyLoss()
-        output = module(*sample_input)
-        loss = criterion(output, module(*sample_input))
-        optimizer.zero_grad()
-        loss.backward()
-        optimizer.step()
-
-        module = torch.ao.quantization.quantize_pt2e.convert_pt2e(module)
+        prepared = self.get_prepared_qat_module(module, sample_input)
+        module = self.get_converted_sgd_trained_module(module, prepared, sample_input)
         self.lower_module_and_test_output(module, sample_input)
 
     def test_qnn_backend_log_softmax(self):
