+++ examples/qualcomm/scripts/export_example.py
@@ -4,10 +4,7 @@ import copy
 
 import torch
 from executorch.backends.qualcomm.partition.qnn_partitioner import QnnPartitioner
-from executorch.backends.qualcomm.quantizer.quantizer import (
-    get_default_8bit_qnn_ptq_config,
-    QnnQuantizer,
-)
+from executorch.backends.qualcomm.quantizer.quantizer import QnnQuantizer
 from executorch.backends.qualcomm.serialization.qnn_compile_spec_schema import (
     QcomChipset,
 )
@@ -64,8 +61,6 @@ def main() -> None:
 
     # Get quantizer
     quantizer = QnnQuantizer()
-    quant_config = get_default_8bit_qnn_ptq_config()
-    quantizer.set_bit8_op_quant_config(quant_config)
 
     # Typical pytorch 2.0 quantization flow
     m = torch.export.export(model.eval(), example_inputs).module()
